Add start tests with library panel type downlods to LibraryCoordinatorTests
